backend/requirements.txt
adafruit-ampy==1.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
av==14.4.0
bitarray==3.0.0
bitstring==4.3.0
boto3==1.38.44
botocore==1.38.44
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
defusedxml==0.8.0rc2
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-import-export==4.3.8
django-mathfilters==1.0.0
django-storages==1.14.6
djangorestframework==3.16.0
dnspython==2.7.0
docstring_parser==0.16
drf-yasg==1.21.10
ecdsa==0.19.0
esptool==4.8.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.6.0
fsspec==2024.12.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.101.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.24.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
inflection==0.5.1
intelhex==2.3.0
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpfshell==0.9.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pygame==2.5.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMuPDF==1.24.14
pyOpenSSL==25.1.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytransform3d==3.14.0
pytz==2025.2
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
sympy==1.13.1
tablib==3.8.0
tenacity==8.5.0
torch==2.5.1
torchvision==0.20.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
your-package==1.0
